+++ datagateway_api/src/resources/search_api_endpoints.py
@@ -3,6 +3,7 @@ import logging
 from flask_restful import Resource
 
 from datagateway_api.src.common.helpers import get_filters_from_query_string
+from datagateway_api.src.search_api.filters import SearchAPIScoringFilter
 from datagateway_api.src.search_api.helpers import (
     get_count,
     get_files,
@@ -11,6 +12,7 @@ from datagateway_api.src.search_api.helpers import (
     get_with_pid,
     search_api_error_handling,
 )
+from datagateway_api.src.search_api.search_scoring import SearchScoring
 
 log = logging.getLogger()
 
@@ -19,7 +21,7 @@ def get_search_endpoint(entity_name):
     """
     Given an entity name, generate a flask_restful `Resource` class. In
     `create_api_endpoints()`, these generated classes are registered with the API e.g.
-    `api.add_resource(get_search_endpoint("Dataset"), "/datasets")`
+    `api.add_resource(get_search_endpoint("Dataset"), "/Datasets")`
 
     :param entity_name: Name of the entity
     :type entity_name: :class:`str`
@@ -30,8 +32,20 @@ def get_search_endpoint(entity_name):
         @search_api_error_handling
         def get(self):
             filters = get_filters_from_query_string("search_api", entity_name)
-            log.debug("Filters: %s", filters)
-            return get_search(entity_name, filters), 200
+            results = get_search(entity_name, filters)
+            scoring_filter = next(
+                (
+                    filter_
+                    for filter_ in filters
+                    if isinstance(filter_, SearchAPIScoringFilter)
+                ),
+                None,
+            )
+            if scoring_filter:
+                scores = SearchScoring.get_score(scoring_filter.value)
+                results = SearchScoring.add_scores_to_results(results, scores)
+
+            return results, 200
 
         get.__doc__ = f"""
             ---
@@ -66,7 +80,7 @@ def get_single_endpoint(entity_name):
     """
     Given an entity name, generate a flask_restful `Resource` class. In
     `create_api_endpoints()`, these generated classes are registered with the API e.g.
-    `api.add_resource(get_single_endpoint("Dataset"), "/datasets/<string:pid>")`
+    `api.add_resource(get_single_endpoint("Dataset"), "/Datasets/<string:pid>")`
 
     :param entity_name: Name of the entity
     :type entity_name: :class:`str`
@@ -116,7 +130,7 @@ def get_number_count_endpoint(entity_name):
     """
     Given an entity name, generate a flask_restful `Resource` class. In
     `create_api_endpoints()`, these generated classes are registered with the API e.g.
-    `api.add_resource(get_number_count_endpoint("Dataset"), "/datasets/count")`
+    `api.add_resource(get_number_count_endpoint("Dataset"), "/Datasets/count")`
 
     :param entity_name: Name of the entity
     :type entity_name: :class:`str`
@@ -161,7 +175,7 @@ def get_files_endpoint(entity_name):
     """
     Given an entity name, generate a flask_restful `Resource` class. In
     `create_api_endpoints()`, these generated classes are registered with the API e.g.
-    `api.add_resource(get_files_endpoint("Dataset"), "/datasets/<string:pid>/files")`
+    `api.add_resource(get_files_endpoint("Dataset"), "/Datasets/<string:pid>/files")`
 
     :param entity_name: Name of the entity
     :type entity_name: :class:`str`
@@ -217,7 +231,7 @@ def get_number_count_files_endpoint(entity_name):
     Given an entity name, generate a flask_restful `Resource` class. In
     `create_api_endpoints()`, these generated classes are registered with the API e.g.
     `api.add_resource(get_number_count_files_endpoint("Dataset"),
-    "/datasets<string:pid>/files/count")`
+    "/Datasets<string:pid>/files/count")`
 
     :param entity_name: Name of the entity
     :type entity_name: :class:`str`